mm: Refactor background writeback against synchronous
Often, we want to start background writeback a bit before synchronous
writeback to avoid ever reaching a blocking writeback, but not so early
that we are constantly writing back dirty pages while not even coming
close to our dirty writeback threshold.

Lock the dirty background tunables to a fixed ratio against the dirty
ratio tunables.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I9d6569ef0d6413af368fe260aa9a026298784b6d
Signed-off-by: negrroo <mohammedaelnaggar1@gmail.com>